+++ Workload/app/items/PackageInstallerItem/Readme.md
@@ -203,3 +203,26 @@ Extend the component library for specialized use cases.
 ### Package Sources
 
 Extend `PackageRegistry` to support external repositories and version management.
+
+## Recent Improvements
+
+### Documentation Updates ✅
+
+- Comprehensive documentation added for all deployment strategies
+- Added detailed class and method documentation
+- Fixed documentation typos and inconsistencies
+- Updated API documentation with usage examples
+
+### Code Quality Fixes ✅
+
+- Fixed typo in `DeployedItem.itemDefinitionName` interface property
+- Improved error handling in deployment strategies
+- Enhanced type safety across deployment components
+- Updated method signatures for better clarity
+
+### OneLake Storage Client Fixes ✅
+
+- Fixed `readFileAsBase64` method to properly handle binary content
+- Corrected URL formatting for OneLake DFS API calls
+- Improved binary data encoding and decoding processes
+- Enhanced error handling for OneLake operations
