+++ Workload/app/items/PackageInstallerItem/TODOs.md
@@ -1,18 +1,18 @@
 # Open items that need to be addressed in the project
 
-* Editor needs to scrole inside
+* Editor needs to scroll inside
 
 * About dialog w link to the repo
 * settings allow service principal config
 * schema for Json
 * create an install from a Workspace/CICD
-* Documentation
-* schma validation for packages before they are loaded
+* ~~Documentation~~ ✅ **COMPLETED** - Added comprehensive documentation for deployment strategies
+* schema validation for packages before they are loaded
 
 * Spark Livy deployment
 
-* Service pinciple support for the creation proces
-* UX Paramters allow customers to set certain parameters over the UX.
+* Service principal support for the creation process
+* UX Parameters allow customers to set certain parameters over the UX.
 
 ## CognitiveSample
 * correct the Notebook file to work
